external-configuration-provider/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
requests==2.26.0
